Converted from doubles to decimals

+++ Domain.Tests/CalculatePriceTest.cs
@@ -8,7 +8,7 @@ public class CalculatePriceTest
     public void OrderPriceShouldContainDiscountWhenOrderHasStudentTicket()
     {
         Movie movie = new Movie("Testmovie");
-        MovieScreening screening = new MovieScreening(movie, DateTime.Parse("13:00 01-01-2023"),10.00);
+        MovieScreening screening = new MovieScreening(movie, DateTime.Parse("13:00 01-01-2023"),10.00m);
         MovieTicket ticket1 = new MovieTicket(screening, false, 1, 1, true);
         MovieTicket ticket2 = new MovieTicket(screening, false, 1, 2, false);
 
@@ -16,14 +16,14 @@ public class CalculatePriceTest
         order.AddSeatReservation(ticket1);
         order.AddSeatReservation(ticket2);
 
-        Assert.Equal(10.00, order.CalculatePrice());
+        Assert.Equal(10.00m, order.CalculatePrice());
     }
 
     [Fact]
     public void OrderPriceShouldContainDiscountWhenOrderIsNotInWeekend()
     {
         Movie movie = new Movie("Testmovie");
-        MovieScreening screening = new MovieScreening(movie, DateTime.Parse("13:00 03-01-2023"),10.00);
+        MovieScreening screening = new MovieScreening(movie, DateTime.Parse("13:00 03-01-2023"),10.00m);
         MovieTicket ticket1 = new MovieTicket(screening, false, 1, 1, false);
         MovieTicket ticket2 = new MovieTicket(screening, false, 1, 2, false);
 
@@ -31,16 +31,16 @@ public class CalculatePriceTest
         order.AddSeatReservation(ticket1);
         order.AddSeatReservation(ticket2);
 
-        double sut = order.CalculatePrice();
+        decimal sut = order.CalculatePrice();
 
-        Assert.Equal(10.00, sut);
+        Assert.Equal(10.00m, sut);
     }
 
         [Fact]
     public void OrderPriceShouldContainGroupDiscountWhenOrderIsLargerThanOrEqualTo6Tickets()
     {
         Movie movie = new Movie("Testmovie");
-        MovieScreening screening = new MovieScreening(movie, DateTime.Parse("13:00 01-01-2023"),10.00);
+        MovieScreening screening = new MovieScreening(movie, DateTime.Parse("13:00 01-01-2023"),10.00m);
         MovieTicket ticket1 = new MovieTicket(screening, false, 1, 1, false);
         MovieTicket ticket2 = new MovieTicket(screening, false, 1, 2, false);
         MovieTicket ticket3 = new MovieTicket(screening, false, 1, 3, false);
@@ -56,16 +56,16 @@ public class CalculatePriceTest
         order.AddSeatReservation(ticket5);
         order.AddSeatReservation(ticket6);
 
-        double sut = order.CalculatePrice();
+        decimal sut = order.CalculatePrice();
         
-        Assert.Equal(54.00, sut);
+        Assert.Equal(54.00m, sut);
     }
 
     [Fact]
     public void OrderPriceShouldNotContainGroupDiscountWhenOrderIsLessThan6Tickets()
     {
         Movie movie = new Movie("Testmovie");
-        MovieScreening screening = new MovieScreening(movie, DateTime.Parse("13:00 01-01-2023"),10.00);
+        MovieScreening screening = new MovieScreening(movie, DateTime.Parse("13:00 01-01-2023"),10.00m);
         MovieTicket ticket1 = new MovieTicket(screening, false, 1, 1, false);
         MovieTicket ticket2 = new MovieTicket(screening, false, 1, 2, false);
         MovieTicket ticket3 = new MovieTicket(screening, false, 1, 3, false);
@@ -79,54 +79,54 @@ public class CalculatePriceTest
         order.AddSeatReservation(ticket4);
         order.AddSeatReservation(ticket5);
 
-        double sut = order.CalculatePrice();
+        decimal sut = order.CalculatePrice();
         
-        Assert.Equal(50.00, sut);
+        Assert.Equal(50.00m, sut);
     }
 
     [Fact]
     public void OrderPriceShouldHave2EuroPremiumForPremiumSeatsWhenTicketHasStudentDiscount()
     {
         Movie movie = new Movie("Testmovie");
-        MovieScreening screening = new MovieScreening(movie, DateTime.Parse("13:00 01-01-2023"),10.00);
+        MovieScreening screening = new MovieScreening(movie, DateTime.Parse("13:00 01-01-2023"),10.00m);
         MovieTicket ticket1 = new MovieTicket(screening, true, 1, 1, true);
 
         Order order = new Order(1);
         order.AddSeatReservation(ticket1);
 
-        double sut = order.CalculatePrice();
+        decimal sut = order.CalculatePrice();
         
-        Assert.Equal(12.00, sut);
+        Assert.Equal(12.00m, sut);
     }
 
     [Fact]
     public void OrderPriceShouldHave3EuroPremiumForPremiumSeatsWhenTicketDoesNotHaveStudentDiscount()
     {
         Movie movie = new Movie("Testmovie");
-        MovieScreening screening = new MovieScreening(movie, DateTime.Parse("13:00 01-01-2023"),10.00);
+        MovieScreening screening = new MovieScreening(movie, DateTime.Parse("13:00 01-01-2023"),10.00m);
         MovieTicket ticket1 = new MovieTicket(screening, true, 1, 1, false);
 
         Order order = new Order(1);
         order.AddSeatReservation(ticket1);
 
-        double sut = order.CalculatePrice();
+        decimal sut = order.CalculatePrice();
         
-        Assert.Equal(13.00, sut);
+        Assert.Equal(13.00m, sut);
     }
 
     [Fact]
     public void OrderPriceShouldCost0EurosForPremiumSeatWhenTicketIsFree()
     {
         Movie movie = new Movie("Testmovie");
-        MovieScreening screening = new MovieScreening(movie, DateTime.Parse("13:00 01-01-2023"),10.00);
+        MovieScreening screening = new MovieScreening(movie, DateTime.Parse("13:00 01-01-2023"),10.00m);
         MovieTicket ticket1 = new MovieTicket(screening, true, 1, 1, true);
         MovieTicket ticket2 = new MovieTicket(screening, true, 1, 1, false);
         
         Order order = new Order(1);
         order.AddSeatReservation(ticket1);
 
-        double sut = order.CalculatePrice();
+        decimal sut = order.CalculatePrice();
         
-        Assert.Equal(12.00, sut);
+        Assert.Equal(12.00m, sut);
     }
 }
